pdf-translator/requirements.txt
certifi==2025.1.31
chardet==3.0.4
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2025.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
packaging==24.2
pdf2image==1.17.0
pillow==11.1.0
PyPDF2==3.0.1
pytesseract==0.3.13
reportlab==4.3.0
rfc3986==1.5.0
sniffio==1.3.1
